transceiver: fix calculation of training sequence correlation for 2 sps
The correlation starting point for normal burst training sequence
calculation should be a scaled value of the same symbol regardless
of the samples-per-symbol used. Use of 2 samples-per-symbol double
the index values, but yields the following outputs, which results
in a late time-of-arrival value at the output of the correlation.
Modifiy the parameter calculation according to match expected
input.

1 sps parameters

maxTOA = 3
spanTOA = 5;
startIx = 61;
endIx = 87;
windowLen = 26;
corrLen = 7;

2 sps parameters (errant case)

maxTOA = 6;
spanTOA = 10;
startIx = 112;
endIx = 184;
windowLen = 72;
corrLen =13;

Signed-off-by: Thomas Tsou <ttsou@vt.edu>